+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### July 22, 2026
+- 713400: Optimalized Web Portal Dashboard component widgets coupling
+
 ### July 13, 2026
 - 713184: Optimizing request history and pending request loading (Need API update).
 - 724361: Fixing pending attestation grouping.